+++ primitives/native-token-management/src/lib.rs
@@ -52,7 +52,7 @@ impl MainChainScripts {
 
 sp_api::decl_runtime_apis! {
 	pub trait NativeTokenManagementApi {
-		fn get_main_chain_scripts() -> MainChainScripts;
+		fn get_main_chain_scripts() -> Option<MainChainScripts>;
 	}
 }
 
@@ -92,7 +92,7 @@ mod inherent_provider {
 	use std::sync::Arc;
 
 	pub struct NativeTokenManagementInherentDataProvider {
-		pub token_amount: NativeTokenAmount,
+		pub token_amount: Option<NativeTokenAmount>,
 	}
 
 	#[derive(thiserror::Error, sp_runtime::RuntimeDebug)]
@@ -119,9 +119,12 @@ mod inherent_provider {
 			C::Api: NativeTokenManagementApi<Block>,
 		{
 			let api = client.runtime_api();
-			let scripts = api
+			let Some(scripts) = api
 				.get_main_chain_scripts(parent_hash)
-				.map_err(IDPCreationError::GetMainChainScriptsError)?;
+				.map_err(IDPCreationError::GetMainChainScriptsError)?
+			else {
+				return Ok(Self { token_amount: None });
+			};
 			let parent_mc_hash: Option<McBlockHash> =
 				get_mc_hash_for_block(client.as_ref(), parent_hash)
 					.map_err(IDPCreationError::McHashError)?;
@@ -135,6 +138,8 @@ mod inherent_provider {
 				)
 				.await?;
 
+			let token_amount = Some(token_amount).filter(|amount| amount.0 > 0);
+
 			Ok(Self { token_amount })
 		}
 	}
@@ -145,10 +150,11 @@ mod inherent_provider {
 			&self,
 			inherent_data: &mut InherentData,
 		) -> Result<(), sp_inherents::Error> {
-			inherent_data.put_data(
-				INHERENT_IDENTIFIER,
-				&TokenTransferData { token_amount: self.token_amount.clone() },
-			)
+			if let Some(token_amount) = self.token_amount {
+				inherent_data.put_data(INHERENT_IDENTIFIER, &TokenTransferData { token_amount })
+			} else {
+				Ok(())
+			}
 		}
 
 		async fn try_handle_error(