+++ primitives/native-token-management/src/tests/mod.rs
@@ -4,6 +4,7 @@ pub(crate) mod runtime_api_mock;
 mod inherent_provider {
 	use super::runtime_api_mock::*;
 	use crate::inherent_provider::*;
+	use crate::MainChainScripts;
 	use crate::INHERENT_IDENTIFIER;
 	use main_chain_follower_api::mock_services::MockNativeTokenDataSource;
 	use sidechain_domain::*;
@@ -25,7 +26,8 @@ mod inherent_provider {
 
 		let data_source =
 			create_data_source(parent_mc_hash.clone(), mc_hash.clone(), total_transfered);
-		let client = create_client(parent_hash, parent_mc_hash, parent_number);
+		let main_chain_scripts = Some(MainChainScripts::default());
+		let client = create_client(parent_hash, parent_mc_hash, parent_number, main_chain_scripts);
 
 		let inherent_provider = NativeTokenManagementInherentDataProvider::new(
 			client,
@@ -36,7 +38,7 @@ mod inherent_provider {
 		.await
 		.expect("Should not fail");
 
-		assert_eq!(inherent_provider.token_amount.0, total_transfered)
+		assert_eq!(inherent_provider.token_amount, Some(NativeTokenAmount(total_transfered)))
 	}
 
 	#[tokio::test]
@@ -50,7 +52,8 @@ mod inherent_provider {
 
 		let data_source =
 			create_data_source(parent_mc_hash.clone(), mc_hash.clone(), total_transfered);
-		let client = create_client(parent_hash, parent_mc_hash, parent_number);
+		let main_chain_scripts = Some(MainChainScripts::default());
+		let client = create_client(parent_hash, parent_mc_hash, parent_number, main_chain_scripts);
 
 		let inherent_provider = NativeTokenManagementInherentDataProvider::new(
 			client,
@@ -61,11 +64,11 @@ mod inherent_provider {
 		.await
 		.expect("Should not fail");
 
-		assert_eq!(inherent_provider.token_amount.0, total_transfered)
+		assert_eq!(inherent_provider.token_amount, Some(NativeTokenAmount(total_transfered)))
 	}
 
 	#[tokio::test]
-	async fn defaults_to_zero_when_no_data() {
+	async fn defaults_to_none_when_no_data() {
 		let parent_number = 1;
 
 		let mc_hash = McBlockHash([1; 32]);
@@ -73,7 +76,8 @@ mod inherent_provider {
 		let parent_mc_hash = Some(McBlockHash([3; 32]));
 
 		let data_source = MockNativeTokenDataSource::new([].into());
-		let client = create_client(parent_hash, parent_mc_hash, parent_number);
+		let main_chain_scripts = Some(MainChainScripts::default());
+		let client = create_client(parent_hash, parent_mc_hash, parent_number, main_chain_scripts);
 
 		let inherent_provider = NativeTokenManagementInherentDataProvider::new(
 			client,
@@ -84,7 +88,33 @@ mod inherent_provider {
 		.await
 		.expect("Should not fail");
 
-		assert_eq!(inherent_provider.token_amount.0, 0)
+		assert_eq!(inherent_provider.token_amount, None);
+	}
+
+	#[tokio::test]
+	async fn defaults_to_none_when_scripts_are_unset() {
+		let parent_number = 1; // not genesis
+
+		let mc_hash = McBlockHash([1; 32]);
+		let parent_hash = Hash::from([2; 32]);
+		let parent_mc_hash = Some(McBlockHash([3; 32]));
+		let total_transfered = 103;
+
+		let data_source =
+			create_data_source(parent_mc_hash.clone(), mc_hash.clone(), total_transfered);
+		let main_chain_scripts = None;
+		let client = create_client(parent_hash, parent_mc_hash, parent_number, main_chain_scripts);
+
+		let inherent_provider = NativeTokenManagementInherentDataProvider::new(
+			client,
+			&data_source,
+			mc_hash,
+			parent_hash,
+		)
+		.await
+		.expect("Should not fail");
+
+		assert_eq!(inherent_provider.token_amount, None)
 	}
 
 	#[tokio::test]
@@ -94,7 +124,7 @@ mod inherent_provider {
 		let mut inherent_data = InherentData::new();
 
 		let inherent_provider = NativeTokenManagementInherentDataProvider {
-			token_amount: NativeTokenAmount(token_amount),
+			token_amount: Some(NativeTokenAmount(token_amount)),
 		};
 
 		inherent_provider.provide_inherent_data(&mut inherent_data).await.unwrap();
@@ -122,6 +152,7 @@ mod inherent_provider {
 		parent_hash: Hash,
 		parent_mc_hash: Option<McBlockHash>,
 		parent_number: u32,
+		main_chain_scripts: Option<MainChainScripts>,
 	) -> Arc<TestApi> {
 		Arc::new(TestApi {
 			headers: [(
@@ -143,6 +174,7 @@ mod inherent_provider {
 				},
 			)]
 			.into(),
+			main_chain_scripts,
 		})
 	}
 }